pre-rq rebalance checkpoint

+++ commands/trackinfo.js
@@ -62,7 +62,9 @@ module.exports = {
                     errors: ["time"]
                 })
                     .then(collected => {
-						collected.first().delete();
+						if (message.channel.type === "text") {
+							collected.first().delete();
+						}
                         if (isNaN(collected.first().content) || parseInt(collected.first().content) > searchResults.length || parseInt(collected.first().content) < 1) {
 							message.client.execList.splice(message.client.execList.indexOf(message.author.id), 1);
                             const errorMessage = new Discord.MessageEmbed()
